+++ src/dcos_e2e/backends/_aws/__init__.py
@@ -241,7 +241,7 @@ class AWSCluster(ClusterManager):
         """
         return dict(self.launcher.config['dcos_config'])
 
-    def install_dcos_from_url(
+    def install_dcos_from_url_with_bootstrap_node(
         self,
         build_artifact: str,
         dcos_config: Dict[str, Any],